+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/InputMonitoring/ExternalTracker/ExternalTrackerSettingReceiver.cs
@@ -27,11 +27,6 @@ namespace Baku.VMagicMirror
                 VmmCommands.ExTrackerSetFaceSwitchSetting,
                 c => dataSource.SetFaceSwitchSetting(c.Content)
                 );
-            
-            receiver.AssignCommandHandler(
-                VmmCommands.DisableFaceTrackingHorizontalFlip,
-                c => dataSource.SetFaceHorizontalFlipDisable(c.ToBoolean())
-                );
         }
     }
 }