+++ WPF/VMagicMirrorConfig/Model/Entity/GameInputSetting.cs
@@ -1,5 +1,17 @@
-﻿namespace Baku.VMagicMirrorConfig
+﻿using System;
+using System.Windows.Input;
+
+namespace Baku.VMagicMirrorConfig
 {
+    public enum GameInputLocomotionStyle
+    {
+        //スティックの向きはアバターの移動方向そのものであり、かつアバターの体は同じ方向を向き続ける
+        FirstPerson,
+        //スティックの向きにアバターの体が向き直る。左右や奥にスティックを倒した場合、アバターの顔は見えない事もある。
+        ThirdPerson,
+        //横スクロール用の3人称挙動で、左か右のいずれかにしか向かない
+        SideView2D,
+    }
 
     public enum GameInputStickAction
     {
@@ -92,6 +104,7 @@
         public bool UseShiftRun { get; set; } = true;
         public bool UseSpaceJump { get; set; } = true;
 
+        //NOTE: これらのxxxKeyCodeにはSystem.Windows.Forms.KeyをToStringしたものが入る。カラの場合、アサインが無いことを示す
         public string JumpKeyCode { get; set; } = "Space";
         public string RunKeyCode { get; set; } = "Shift";
         public string CrouchKeyCode { get; set; } = "C";
@@ -100,6 +113,48 @@
         public string PunchKeyCode { get; set; } = "";
 
         public static GameInputKeyboardKeyAssign LoadDefault() => new();
+
+        //Unityに投げつける用に前処理したデータを生成する
+        public GameInputKeyboardKeyAssign GetKeyCodeTranslatedData()
+        {
+            var result = new GameInputKeyboardKeyAssign()
+            {
+                UseMouseLookAround = UseMouseLookAround,
+                LeftClick = LeftClick,
+                RightClick = RightClick,
+                MiddleClick = MiddleClick,
+                UseWasdMove = UseWasdMove,
+                UseArrowKeyMove = UseArrowKeyMove,
+                UseShiftRun = UseShiftRun,
+                UseSpaceJump = UseSpaceJump,
+            };
+
+            result.JumpKeyCode = TranslateKeyCode(JumpKeyCode);
+            result.RunKeyCode = TranslateKeyCode(RunKeyCode);
+            result.CrouchKeyCode = TranslateKeyCode(CrouchKeyCode);
+            result.TriggerKeyCode = TranslateKeyCode(TriggerKeyCode);
+            result.PunchKeyCode = TranslateKeyCode(PunchKeyCode);
+
+            return result;
+        }
+
+        private static string TranslateKeyCode(string wpfKey)
+        {
+            if (string.IsNullOrEmpty(wpfKey))
+            {
+                return "";
+            }
+
+            if (!Enum.TryParse<Key>(wpfKey, out var key))
+            {
+                return "";
+            }
+
+            //渋い気もするが、このkの整数値をテキストで書き込む
+            // -> Unity側はこの整数値をWindows.Forms.Keyとして解釈する
+            var k = KeyInterop.VirtualKeyFromKey(key);
+            return k.ToString();
+        }
     }
 
 
@@ -118,6 +173,7 @@
         public bool GamepadEnabled { get; set; } = true;
         public bool KeyboardEnabled { get; set; } = true;
         public bool AlwaysRun { get; set; } = true;
+        public int LocomotionStyleValue { get; set; } = 0;
 
         public GameInputKeyboardKeyAssign KeyboardKeyAssign { get; set; } = new();
         public GameInputGamepadKeyAssign GamepadKeyAssign { get; set; } = new();