+++ WPF/VMagicMirrorConfig/Model/HotKey/HotKeyItem.cs
@@ -8,6 +8,14 @@ namespace Baku.VMagicMirrorConfig
         SetCamera,
         CallWtm,
         ToggleAccessory,
+        SetBodyMotionStyle,
+    }
+
+    public enum HotKeyActionBodyMotionStyle : int
+    {
+        Default = 0,
+        AlwaysHandDown = 1,
+        GameInputLocomotion = 2,        
     }
 
     /// <summary>